remove duplicate definitions of gov.uk form widgets and separate widgets and form fields into more appropriately-named modules
