+++ application/presentation/first_aid/forms/training_details.py
@@ -2,7 +2,7 @@ from django import forms
 
 from application.presentation.utilities import NannyForm
 from datetime import date
-from application.presentation.customfields import CustomSplitDateField
+from application.presentation.form_fields import CustomSplitDateField
 from dateutil.relativedelta import relativedelta
